requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
gunicorn==25.1.0
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
dnspython==2.8.0
SQLAlchemy==2.0.46
aiosqlite==0.22.1
httpx==0.28.1
flake8==7.3.0
sqladmin==0.23.0
PyJWT==2.11.0
itsdangerous==2.2.0
alembic==1.18.4
APScheduler==3.11.2
